Hack device storage for ronin
gaia code uses gonk only APIs (DownloadsAPI) and has some hardcoded
expectation (for instance, it assumes there is a 'sdcard' devicestorage).

What this hack does is using absolute paths instead of relative path in
a deviceStorage context (eg: /a/b/c/d versus b/c/d stored in a deviceStorage
with a moutPoint=/a).

Downloads use XDG app structure so this is enough to be able to download a file
and open it later.