Correctly extract manifest for third-party packaged apps
When not b2gdroid or not running on gonk, the previous logic completely skipped
moving preinstalled apps to their rightful location and extracting manifests.
Therefore, the manifests that we took into account was the update mini-manifest,
missing important stuff like launch_path, web activities configuration etc.

This patch restore the extract logic for ronin, without restoring the move logic
that makes no sense when everything is in one place in the profile.